refactor: update type imports to use specific directories and improve type organization
