+++ src/types/index.d.ts
@@ -13,7 +13,5 @@ export * from './conversations';
 export * from './filters';
 export * from './messages';
 export * from './participant';
-export * from './search';
-export * from './share';
 export * from './soap';
 export * from './utils';
