+++ src/views/search/list/message/search-message-list.tsx
@@ -13,7 +13,7 @@ import { map } from 'lodash';
 import { useParams } from 'react-router-dom';
 
 import { useMultipleSelection } from 'hooks/use-multiple-selection';
-import type { SearchListProps } from 'types/index.d';
+import { SearchListProps } from 'types/search';
 import { MessageShortcutsRegister } from 'views/app/folder-panel/messages/message-shortcuts-register';
 import { MessagesMultipleSelectionActions } from 'views/app/folder-panel/messages/messages-multiple-selection-actions';
 import { SearchMessageListItemWrapper } from 'views/search/list/message/search-message-list-item-wrapper';
